atButtons flat fix 2
atButtons flat fix 2: After thinking on this overnight, I moved the code into the PaintTo but inside a compiler conditional. Logically that's whee it should be, and it does not effect the lazarus version. This all comes up for me from time to time since delphi does not always handle clNone like you think it would. It does sometimes, other times it doesn't.