Update attabs.pas
Added event onTabChanged that is only fired when a tab is changed. This overcomes an issue with onTabClick which had been changed but caused issues in other AT components.. This event does what it say, fires on tab change.

Added property to make a tab visible or not, and makes it more compliant to the regular Lazarus/Delphi components. To hide a tab programmatically you can access the tab store,  or use the function HideTab(index) that also takes care of shifting the tabindex if needed. Same for ShowTab(index).